requirements.txt
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
pyaes==1.6.1
pyasn1==0.5.0
PySocks==1.7.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
spotipy==2.23.0
Telethon==1.30.3
urllib3==2.0.5
yamlordereddictloader==0.4.2
